Improved page select dialog
I really like the preview feature of the page select dialog, but the
slider was too wide for my taste in my netbook screen, so I modified the
dialog to make better use of screen space. I added a text box, so if you
know which page you want, you can type it in directly. The preview is
shown above the box, and the slider is placed vertically next to it.

Author: Nephiel <nephiel@users.sourceforge.net>

git-svn-id: https://svn.code.sf.net/p/mcomix/code@14 ef8490bc-f317-4725-81ed-e95a9aca1bcf