Dynamic background color fix
With MComix 0.90.3, when the "Dynamic background color" preference is
enabled, black and white or grayscale pages are displayed on a gray
background color. This causes the background to stand out instead of
letting the pages fade into it.

In comparison, the old Comix 4.0.4 uses a background color that is
closer to the actual color of the edge of the page.
E.g. if a page has a thick white margin all around, the background
should be way more white than gray, regardless of the amount of black
inside the margin.

This patch changes the function used to get the most used edge color in
image_tools.py to calculate it like the old Comix did.

Author: Nephiel <nephiel@users.sourceforge.net>

git-svn-id: https://svn.code.sf.net/p/mcomix/code@15 ef8490bc-f317-4725-81ed-e95a9aca1bcf