Enable "Go to page" before thumbnails are loaded.
As of MComix svn 26, due to the image preview feature, the "Go to page"
feature is completely disabled until all the thumbnails for the current
file have been loaded.

Case scenario: you're reading a 600-page comic. You close MComix, and
open it again some time later. You were about a third into the comic,
page 200 or so, but you didn't bookmark it, and to jump to that page you
have to wait until all 600 thumbnails have been loaded. That can take a
long time.

This patch makes the "Go to page" feature available as soon as a file is
loaded, just like the rest of the navigation options. The preview image
will show the stock "missing image" icon until all the thumbnails have
been loaded, but the rest of the dialog will be fully functional.

From: Nephiel <nephiel@users.sourceforge.net>

git-svn-id: https://svn.code.sf.net/p/mcomix/code@44 ef8490bc-f317-4725-81ed-e95a9aca1bcf