Only open files actually passed to MComix.
In theory, the user can pass more than one file to MComix at
startup. This was, however, mostly ignored by earlier versions,
and only opened the first file passed. Fix this behaviour
by introducing a new class <FileProvider> that is responsible
for keeping track of files that should be opened.

Now, calling FileHandler.open_file with a list of files instead
of a single file will only add these files to the FileProvider.
If a string is passed, the old behaviour is (hopefully) retained.

git-svn-id: https://svn.code.sf.net/p/mcomix/code@192 ef8490bc-f317-4725-81ed-e95a9aca1bcf