Store recent page even if current page is last in archive.
Previously, the first and last page of an archive would not
be stored in the database. However, storing the last page
would be a sensible indicator that the book has been read
completely, so allow that information to be stored.

In order to prevent that the popup dialog on opening an archive
would bug the user to resume from the last page, return None
from LastPageRead.get_page if the stored page is the last one.

git-svn-id: https://svn.code.sf.net/p/mcomix/code@766 ef8490bc-f317-4725-81ed-e95a9aca1bcf